prioritize connect_signal over add_signal
In newer versions of awesome, add_signal is used to register a new signal
type with an object; in older versions, it was used to connect a handler
to a signal, functionality now provided as connect_signal. To make sure the
correct function is called, use add_signal only if connect_signal does not
exist.